Intermediate commit for perf test.
Complete the perf test and buffer control. Due to the overhead of
deployed job, this would always be slow than the current feed which uses
a single ong-running job.
Considering to move to decoupled intake and connect, and put the feed
records into the parameter list of an expression.
-- Update: add simple modification to add the callOp to the top of
connJob. Next, change to read from intake job directly.

Change-Id: I5c4a8072eb97bdfaba70491d7f9789bd2408b165